feat(cli): implement pagination support
This implements handling API endpoints, which does not return the full set, but
just a slice. This is the case for pagination. In this case, the API fetches
all data in subsequent requests, collects all the data and return the combined
result in a table.

Because not all "List"-type endpoints support pagination, a whitelist is
introduced to determine which list endpoint can handle pagination and which is
not. This whitelist is not ideal, so a rather big refactoring is done in the
backend, which adds pagination to ALL endpoints starting with "List", so we can
get rid of the whitelist in the future.

Changelog: feat