refactor(grpc): integrate gRPC changes
Integrate and add/change all gRPC changes since 6/2024 into the commandline tool.
To support List* endpoints, which do not return a pagination response, a new
action flag is introduced ("no-pagination"). This way, missing gRPC endpoints
can be integrated. For some endpoints, the identifier is not called "Id", so there
is another field ("identifier-key") to define a different key for the identifier.
Further, all definitions where updated and descriptions added.

* New action field "identifier-key"
* New global field "identifier-key"
* New action field "no-pagination"
* All definitions got descriptions
* gRPC changes since 6/2024 are added/changed

Signed-off-by: Aaron Fischer <aaron.fischer@g-portal.cloud>